rename vspin
